# Speed up the package check tests