--- a/test/test_any_automated_install.py
+++ b/test/test_any_automated_install.py
@@ -1104,11 +1104,13 @@ def test_package_manager_has_installer_deps(host):
     output = host.run('''
     source /opt/pihole/basic-install.sh
     package_manager_detect
-    install_dependent_packages ${INSTALLER_DEPS[@]}
+    for i in "${INSTALLER_DEPS[@]}"
+    do
+        ${PKG_CHECK} ${i}
+    done
     ''')
-
-    assert 'No package' not in output.stdout  # centos7 still exits 0...
-    assert output.rc == 0
+    assert 'Unable to locate package' not in output.stdout
+    assert 'No matching Packages to list' not in output.stderr
 
 
 def test_package_manager_has_pihole_deps(host):
@@ -1118,11 +1120,13 @@ def test_package_manager_has_pihole_deps(host):
     source /opt/pihole/basic-install.sh
     package_manager_detect
     select_rpm_php
-    install_dependent_packages ${PIHOLE_DEPS[@]}
+    for i in "${PIHOLE_DEPS[@]}"
+    do
+        ${PKG_CHECK} ${i}
+    done
     ''')
-
-    assert 'No package' not in output.stdout  # centos7 still exits 0...
-    assert output.rc == 0
+    assert 'Unable to locate package' not in output.stdout
+    assert 'No matching Packages to list' not in output.stderr
 
 
 def test_package_manager_has_web_deps(host):
@@ -1132,8 +1136,10 @@ def test_package_manager_has_web_deps(host):
     source /opt/pihole/basic-install.sh
     package_manager_detect
     select_rpm_php
-    install_dependent_packages ${PIHOLE_WEB_DEPS[@]}
+    for i in "${PIHOLE_WEB_DEPS[@]}"
+    do
+        ${PKG_CHECK} ${i}
+    done
     ''')
-
-    assert 'No package' not in output.stdout  # centos7 still exits 0...
-    assert output.rc == 0
+    assert 'Unable to locate package' not in output.stdout
+    assert 'No matching Packages to list' not in output.stderr