Mostly thinking from the point of view of the PHP packages which can take a while to install on the CI (which is why we moved to a pre-built base image on Docker)

Testing locally on my machine for `test_package_manager_has_web_deps`:

before:
```
12.73s call     test_any_automated_install.py::test_package_manager_has_web_deps
0.59s setup    test_any_automated_install.py::test_package_manager_has_web_deps
0.47s teardown test_any_automated_install.py::test_package_manager_has_web_deps
```

after:
```
4.41s call     test_any_automated_install.py::test_package_manager_has_web_deps
0.51s setup    test_any_automated_install.py::test_package_manager_has_web_deps
0.47s teardown test_any_automated_install.py::test_package_manager_has_web_deps
```

Curious to see how that translates on github actions