@PromoFaux 2022-07-30
It's actually possible to further improve this: by passing the whole array to the `PKG_CHECK` command, however this is only effective for `apt`, as `dnf`/`yum` don't display any errors if you pass in a mix of existing and non existing packages

Time on my local machine:
```
2.34s call     test_any_automated_install.py::test_package_manager_has_web_deps
0.58s setup    test_any_automated_install.py::test_package_manager_has_web_deps
0.46s teardown test_any_automated_install.py::test_package_manager_has_web_deps
```

apt:
```
adam@adam-pc:~/pi-hole$ apt-cache show git glit
Package: git
Priority: optional
Section: vcs
[ ... SNIP ... ]

N: Unable to locate package glit
adam@adam-pc:~/pi-hole$ echo $?
0
```

yum doesn't even report it via return code:
```
yum list git glit
Last metadata expiration check: 0:00:26 ago on Sat Jul 30 16:43:54 2022.
Available Packages
git.x86_64                                                                                                                  2.31.1-2.el8                                                                                                                   appstream
[root@b5f20e3970b0 /]# echo $?
0       
```